Move test class, add more verification
